+++ router/rpc_route_from_route.js
@@ -16,8 +16,10 @@ const rpcHopFromHop = require('./rpc_hop_from_hop');
       timeout: <Timeout Block Height Number>
     }]
     mtokens: <Total Fee-Inclusive Millitokens String>
+    [payment]: <Payment Identifier Hex String>
     timeout: <Timeout Block Height Number>
     tokens: <Total Fee-Inclusive Tokens Number>
+    [total_mtokens]: <Total Millitokens String>
   }
 
   @throws
@@ -34,6 +36,11 @@ const rpcHopFromHop = require('./rpc_hop_from_hop');
       fee: <Fee in Tokens Number>
       fee_msat: <Fee in Millitokens Number>
       [pub_key]: <Next Hop Public Key Hex String>
+      tlv_payload: <Has Extra TLV Data Bool>
+      [mpp_record]: {
+        payment_addr: <Payment Identifier Buffer>
+        total_amt_msat: <Total Payment Millitokens Amount String>
+      }
     }]
     total_amt: <Total Tokens Number>
     total_amt_msat: <Route Total Millitokens String>
@@ -43,8 +50,24 @@ const rpcHopFromHop = require('./rpc_hop_from_hop');
   }
 */
 module.exports = args => {
+  const hops = args.hops.map(hop => rpcHopFromHop(hop));
+
+  const payAddress = !args.payment ? null : Buffer.from(args.payment, 'hex');
+
+  // Set the payment identifier and total amount in the TLV payload
+  if (!!args.payment || !!args.total_mtokens) {
+    const finalHopIndex = hops.length - 1
+
+    hops[hops.length - 1].tlv_payload = true;
+
+    hops[hops.length - 1].mpp_record = {
+      payment_addr: payAddress || undefined,
+      total_amt_msat: args.total_mtokens || undefined,
+    };
+  }
+
   return {
-    hops: args.hops.map(hop => rpcHopFromHop(hop)),
+    hops,
     total_amt: args.tokens.toString(),
     total_amt_msat: args.mtokens,
     total_fees: args.fee.toString(),