+++ test/tower_clientrpc-integration/test_get_connected_watchtowers.js
@@ -19,8 +19,22 @@ const {waitForTermination} = require('./../macros');
 
 // Getting connected watchtowers should return watchtowers
 test(`Connect watchtower`, async ({end, equal, match}) => {
-  const control = await spawnLnd({watchers: true});
-  const tower = await spawnLnd({tower: true});
+  let control;
+  let tower;
+
+  try {
+    control = await spawnLnd({watchers: true});
+    tower = await spawnLnd({tower: true});
+  } catch (err) {
+    const [,, failDetails] = err;
+
+    const [failMessage] = failDetails;
+
+    // LND 0.7.1 does not support wtclient
+    if (failMessage === "unknown flag `wtclient.active'") {
+      return end();
+    }
+  }
 
   const info = (await getTowerServerInfo({lnd: tower.lnd})).tower;
   const {lnd} = control;